Felix Update
added a water ("Neutralizer" tag) to reset the mixture in the cooking pot by throwing it in. made an asset for testing.

made a big pink reset button to get new ingredients. it works by manually checking the checkbox ingame. added "OnMouseDown" too, but i cant get it to work, no idea what the issue is.